api/requirements.txt
flask===3.0.0
python-dotenv
networkx
boto3